+++ repository/user_repo.py
@@ -7,8 +7,12 @@ from repository.database.verification import User
 from datetime import datetime
 
 
+def time() -> str:
+    return datetime.today().strftime("%Y-%m-%d %H:%M:%S")
+
+
 class UserRepository(BaseRepository):
-    # unknown - pending - verified - kicked - banned
+    # unknown - pending - verified - kicked - banned - unverified
 
     def add(
         self, discord_id: int, login: str, group: str, code: str,
@@ -25,6 +29,24 @@ class UserRepository(BaseRepository):
         session.query(User).filter(User.discord_id == discord_id).update(
             {User.status: "verified", User.changed: ch}
         )
+        session.commit()
+
+    def update(
+        self,
+        discord_id: int,
+        *,
+        login: str = None,
+        group: str = None,
+        code: str = None,
+        status: str = None,
+    ):
+        """Update user entry"""
+        user = session.query(User).filter(User.discord_id == discord_id).one_or_none()
+        user.login = login or user.login
+        user.group = group or user.group
+        user.code = code or user.code
+        user.status = status or user.status
+        user.changed = time()
         session.commit()
 
     def update_login(self, discord_id: int, login: str):